MachineLearningFiles/LSTM/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
nltk==3.6.2
numpy==1.19.5
numpyencoder==0.3.0
pandas==1.2.4

requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
snowballstemmer==2.1.0
Sphinx==3.5.4

sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0rc3
tensorflow-estimator==2.5.0rc0


